root_names shouldn't be reused
