+++ packages/extractor/src/cache.js
@@ -1,30 +1,4 @@
-// FIXME:
-// Find a better workaround.
-// Current issue: https://github.com/vitest-dev/vitest/issues/6953
-// We want to remove it, for the cross-runtime compatibility.
-import module from "node:module";
-
 import ts from "typescript";
-
-/**
- * @param {string} specifier
- * @returns {URL}
- */
-function get_node_module_filepath(specifier) {
-	if (typeof import.meta.resolve === "function") return new URL(import.meta.resolve(specifier));
-	const require = module.createRequire(import.meta.url);
-	return new URL(`file://${require.resolve(specifier)}`);
-}
-
-/**
- * @returns {string[]}
- */
-function create_default_root_names() {
-	return [
-		//
-		get_node_module_filepath("svelte2tsx/svelte-shims-v4.d.ts").pathname,
-	];
-}
 
 /**
  * @typedef CachedFile
@@ -39,8 +13,6 @@ class Cache {
 	#cached = new Map();
 	/** @type {ts.Program | undefined} */
 	program;
-	/** @type {Set<string>} */
-	root_names = new Set(create_default_root_names());
 	/**
 	 * @param {string} filepath
 	 * @returns {boolean}
